Renamed files consts in constants
